TRECISION: Use ScummVM save file manager and serializer for saved games
Original saved games should be compatible. We can add extra metadata to
the ScummVM-generated saves, which will then be incompatible with the
original interpreter

+++ engines/trecision/nl/classes.cpp
@@ -37,7 +37,7 @@ namespace Trecision {
 #define COUNTERSTEP 6
 
 bool ForceQuit = false;
-int NlVer = 102;
+byte NlVer = 102;
 
 /*-------------------------------------------------------------------------*/
 /*                                DOACTION           					   */
